fix(scan): move ScanResult shaping out of policy, version-bump, thread prior_decision
Three more Codex findings on PR #967, all legitimate:

1. abicheck/policy/AGENTS.md scopes that package to deciding relevance,
   suppression, classification, severity, and gate effect -- never "how is
   it reported" (report/'s job). scan_abort_result_fields's verdict/exit_code
   pairing and {"exit": ...} wrapping is report shape, not a gate decision,
   so it moves out of exit_decision_precedence.py into a new leaf module,
   abicheck/workflows/scan_abort_result.py -- workflows may depend on policy,
   so it still calls resolve_scan_exit_decision for the actual decision.
   Both service_scan.py and scan_engine.py are themselves debt-frozen
   (architecture/debt.yaml), so this couldn't be inlined at either's catch
   sites either; a new module was the only budget-neutral home. Moved the
   matching tests into a new tests/test_scan_abort_result.py alongside it.

2. ScanResult.to_dict()'s report key is covered by SCAN_SCHEMA_VERSION's
   documented additive/breaking bump policy; the new nonempty report.exit
   shape needed a version bump (1.22 -> 1.23) and a history entry, which
   this adds. schemas/__init__.py is also debt-frozen, so a nearby existing
   4-line history paragraph (1.20's entry) is re-wrapped to 3 lines to make
   budget-neutral room, the same technique used for service_scan.py in the
   prior commit.

3. scan_abort_result_fields now accepts an optional prior_decision,
   threading through to resolve_scan_exit_decision's existing parameter of
   the same name -- covers a caller that already has one. Actually
   populating it from scan_engine.py's own *later* _BudgetOverflow raise
   site (the post-compare deadline check) needs a change to that
   exception's constructor to carry the decision across the raise/catch
   boundary, which is real, separate follow-up work recorded in ADR-064
   rather than attempted here, given scan_engine.py's own tight debt budget
   and the risk of a rushed change to that function.

Verified: check_architecture.py (0 errors, real merge-base against
origin/main), mypy, ruff, and the full exit-decision/scan-abort/schema test
surface all pass.
